# APISpec/HETSswagger.yaml
swagger: '2.0'
info:
  version: v1
  title: REST API Documentation for the MOTI Hired Equipment Tracking System (HETS) Application
  description: |
    The Hired Equipment Program is for owners/operators who have a dump truck, bulldozer, backhoe or 
    other piece of equipment they want to hire out to the transportation ministry for day labour and 
    emergency projects. 
    The Hired Equipment Program distributes available work to local equipment owners. The program is 
    based on seniority and is designed to deliver work to registered users fairly and efficiently 
    through the development of local area call-out lists.
  termsOfService: N/A
basePath: /api
paths:
...
  /projects:
    get:
      tags:
        - Project
      consumes: []
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/Project"
    post:
      tags:
        - Project
      consumes:
        - application/json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: item
          in: body
          required: true
          schema:
            $ref: "#/definitions/Project"
      responses:
        201:
          description: Project created
          headers:
            Location:
              description: A link to the Project
              type: string
              format: url
          schema:
            $ref: "#/definitions/Project"  
  /projects/{id}:
    get:
      tags:
        - Project
      consumes: []
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Project"
        404:
          description: Project not found
      parameters:
        - name: id
          in: path
          description: id of Project to fetch
          required: true
          type: integer
    put:
      tags:
        - Project
      consumes:
        - application/json
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Project"
        404:
          description: Project not found
      parameters:
        - name: id
          in: path
          description: id of Project to fetch
          required: true
          type: integer
        - name: item
          in: body
          required: true
          schema:
            $ref: "#/definitions/Project"
  /projects/{id}/delete:
    post:
      tags:
        - Project
      consumes: []
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        200:
          description: OK
        404:
          description: Project not found
      parameters:
        - name: id
          in: path
          description: id of Project to delete
          required: true
          type: integer
  /projects/bulk:
    post:
      tags:
        - Project
      consumes:
        - application/json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: items
          in: body
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/Project"
      responses:
        201:
          description: Project created
...
  /projects/search:
    get:
      summary: Searches Projects
      description:
        Used for the project search page.
      tags:
        - Project
      consumes:
        - application/json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:        
        - name: serviceareas
          in: query
          description: Service Areas (array of id numbers)
          collectionFormat: multi
          required: false
          type: array
          items:
            type: integer       
        - name: project
          in: query
          description: name or partial name for a Project
          required: false
          type: string
        - name: hasRequests
          in: query
          description: if true then only include Projects with active Requests
          required: false
          type: boolean
        - name: hasHires
          in: query
          description: if true then only include Projects with active Rental Agreements
          required: false
          type: boolean
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/ProjectSearchResultViewModel"                            
...
definitions:
  Project:
    type: object
    description: A Provincial Project that my from time to time request equipment under the HETS programme from a Service Area.
    required:
      - id
    properties:
      id:
        format: int32
        type: integer
        description: A system-generated unique identifier for a Project
      localArea:
        $ref: '#/definitions/LocalArea'
        description: The Local Area associated with this Project record.
      provincialProjectNumber:
        type: string
        maxLength: 150
        description: TO BE REVIEWED WITH THE BUSINESS - The Provincial charge code for the equipment hiring related to this project. This will be the same across multiple service areas that provide equipment for the same Project.
      name:
        type: string
        maxLength: 100
        description: A descriptive name for the Project, useful to the HETS Clerk and Project Manager.
      status:
        type: string
        maxLength: 50
        description: The status of the project to determine if it is listed when creating new requests
      information:
        type: string
        maxLength: 2048
        description: Information about the Project needed by the HETS Clerks. Used for capturing varying (project by project) metadata needed to process requests related to the project.
      rentalRequests:
        type: array
        items:
          $ref: '#/definitions/RentalRequest'
      primaryContact:
        $ref: '#/definitions/Contact'
        description: Link to the designated Primary Contact for the Project - usually the Project Manager requesting to hire equipment.
      contacts:
        type: array
        items:
          $ref: '#/definitions/Contact'
      notes:
        type: array
        items:
          $ref: '#/definitions/Note'
      attachments:
        type: array
        items:
          $ref: '#/definitions/Attachment'
      history:
        type: array
        items:
          $ref: '#/definitions/History'
  RentalRequest:
    type: object
    description: A request from a Project for one or more of a type of equipment from a specific Local Area.
    required:
      - id
    properties:
      id:
        format: int32
        type: integer
        description: A system-generated unique identifier for a Request
      project:
        $ref: '#/definitions/Project'
      localArea:
        $ref: '#/definitions/LocalArea'
      status:
        type: string
        maxLength: 50
        description: The status of the Rental Request - whether it in progress, completed or was cancelled.
      equipmentType:
        $ref: '#/definitions/EquipmentType'
      equipmentCount:
        format: int32
        type: integer
        description: The number of pieces of the equipment type wanted for hire as part of this request.
      expectedHours:
        format: int32
        type: integer
        description: The expected number of rental hours for each piece equipment hired against this request, as provided by the Project Manager making the request. 
      expectedStartDate:
        type: string
        format: date-time
        description: The expected start date of each piece of equipment hired against this request, as provided by the Project Manager making the request.
      expectedEndDate:
        type: string
        format: date-time
        description: The expected end date of each piece of equipment hired against this request, as provided by the Project Manager making the request.
      firstOnRotationList:
        $ref: '#/definitions/Equipment'
        description: The first piece of equipment on the rotation list at the time of the creation of the request.
      notes:
        type: array
        items:
          $ref: '#/definitions/Note'
      attachments:
        type: array
        items:
          $ref: '#/definitions/RentalRequestAttachment'
      history:
        type: array
        items:
          $ref: '#/definitions/History'
      rentalRequestRotationList:
        type: array
        items:
          $ref: '#/definitions/RentalRequestRotationList'
  RentalRequestAttachment:
    type: object
    description: Attachments that are required as part of the Rental Requests
    required:
      - id
      - attachment
    properties:
      id:
        format: int32
        type: integer
        description: A system-generated unique identifier for a RentalRequestAttachment
      rentalRequest:
        $ref: '#/definitions/RentalRequest'
      attachment:
        type: string
        maxLength: 150
        description: The name&#x2F;type attachment needed as part of the fulfillment of the request
  RentalRequestRotationList:
    type: object
    description: An eligible piece of equipment for a request and a tracking of the hire offer and response process related to a request for that piece of equipment. Includes a link from the equipment to a Rental Agreement if the equipment was hired to satisfy a part of the request.
    required:
      - id
    properties:
      id:
        format: int32
        type: integer
        description: A system-generated unique identifier for a RequestRotationList
      rentalRequest:
        $ref: '#/definitions/RentalRequest'
      rotationListSortOrder:
        format: int32
        type: integer
        description: The sort order of the piece of equipment on the rotaton list at the time the request was created. This is the order the equipment will be offered the available work.
      equipment:
        $ref: '#/definitions/Equipment'
      rentalAgreement:
        $ref: '#/definitions/RentalAgreement'
        description: The rental agreement (if any) created for an accepted hire offer.
      isForceHire:
        type: boolean
        description: True if the HETS Clerk designated the hire of this equipment as being a Forced Hire. A Force Hire implies that the usual seniority rules for hiring are bypassed because of special circumstances related to the hire - e.g. a the hire requires an attachment only one piece of equipment has.
      wasAsked:
        type: boolean
        description: True if the HETS Clerk contacted the equipment owner and asked to hire the piece of equipment.
      askedDateTime:
        type: string
        format: date-time
        description: The Date-Time the HETS clerk contacted the equipment owner and asked to hire the piece of equipment.
      offerResponse:
        type: string
        description: The response to the offer to hire. Null prior to receiving a response; a string after with the response - likely just Yes or No
      offerRefusalReason:
        type: string
        maxLength: 50
        description: The reason why the user refused the offer based on a selection of values from the UI.
      offerResponseDatetime:
        type: string
        format: date-time
        description: The date and time the final response to the offer was established.
      offerResponseNote:
        type: string
        maxLength: 2048
        description: A note entered about the response to the offer from the equipment owner about the offer. Usually used when the offer is a &quot;No&quot; or &quot;Force Hire&quot;.
      note:
        type: string
        maxLength: 2048
        description: An optional general note about the offer.
...
  Equipment:
    type: object
    description: A piece of equipment in the HETS system. Each piece of equipment is of a specific equipment type, owned by an Owner, and is within a Local Area.
    required:
      - id
    properties:
      id:
        format: int32
        type: integer
        description: A system-generated unique identifier for a Equipment
      localArea:
        $ref: '#/definitions/LocalArea'
      equipmentType:
        $ref: '#/definitions/EquipmentType'
      owner:
        $ref: '#/definitions/Owner'
      equipmentCode:
        type: string
        maxLength: 25
        description: A human-visible unique code for the piece of equipment, referenced for convenience by the system users - HETS Clerks and Equipment Owners. Generated at record creation time based on the unique Owner prefix (e.g. EDW) and a zero-filled unique number - resulting in a code like EDW-0083.
      status:
        type: string
        maxLength: 50
        description: The current status of the equipment in a UI-controlled string. Initial values are Pending, Approved and Archived, but other values may be added.
      receivedDate:
        type: string
        format: date-time
        description: The date the piece of equipment was first received and recorded in HETS.
      approvedDate:
        type: string
        format: date-time
        description: The date the piece of equipment was first approved in HETS. Part of the seniority calculation for a piece of equipment is based on this date.
      lastVerifiedDate:
        type: string
        format: date-time
        description: The date the equipment was last verified by the HETS Clerk as being still in service in the Local Area and available for the HETS Programme.
      isInformationUpdateNeeded:
        type: boolean
        description: Set true if a need to update the information&#x2F;status of the equipment is needed. Used during the processing of a request when an update is noted, but the Clerk does not have time to make the update.
      informationUpdateNeededReason:
        type: string
        maxLength: 2048
        description: A note about why the needed information&#x2F;status update that is needed about the equipment.
      licencePlate:
        type: string
        maxLength: 20
        description: The licence plate (if any) of the piece of equipment, as entered by the HETS Clerk.
      make:
        type: string
        maxLength: 50
        description: The make of the piece of equipment, as provided by the Equipment Owner.
      model:
        type: string
        maxLength: 50
        description: The model of the piece of equipment, as provided by the Equipment Owner.
      year:
        type: string
        maxLength: 15
        description: The model year of the piece of equipment, as provided by the Equipment Owner.
      type:
        type: string
        maxLength: 255
        description: TO BE REVIEWED WITH THE BUSINESS - WHAT IS THIS?
      operator:
        type: string
        maxLength: 255
        description: TO BE REVIEWED WITH THE BUSINESS - WHAT IS THIS?
      payRate:
        type: number
        format: float
        description: TO BE REVIEWED WITH THE BUSINESS - WHAT IS THIS?
      refuseRate:
        type: string
        maxLength: 255
        description: TO BE REVIEWED WITH THE BUSINESS - WHAT IS THIS?
      serialNumber:
        type: string
        maxLength: 100
        description: The serial number of the piece of equipment as provided by the Equipment Owner. Used to detect and reconcile pieces of equipment moved between Local Areas. Duplicate serial numbers are flagged in the system but permitted. The duplicates are flagged in the UI until the HETS Clerks reconcile the differences - either correcting the serial number or archiving a piece of equipment moved to a new local area.
      size:
        type: string
        maxLength: 128
        description: The size of the piece of equipment, as provided by the Equipment Owner.
      toDate:
        type: string
        format: date-time
        description: TO BE REVIEWED WITH THE BUSINESS - WHAT IS THIS?
      blockNumber:
        type: number
        format: float
        description: The current block number for the piece of equipment as calculated by the Seniority Algorthm for this equipment type in the local area. As currently defined y the business  - 1, 2 or Open
      seniority:
        type: number
        format: float
        description: The current seniority calculation result for this piece of equipment. The calculation is based on the &quot;numYears&quot; of service + average hours of service over the last three fiscal years - as stored in the related fields (serviceHoursLastYear, serviceHoursTwoYearsAgo serviceHoursThreeYearsAgo).
      numberInBlock:
        format: int32
        type: integer
        description: The number in the block of the piece of equipment so that it can be displayed to the user where it will be useful. This saves the user from having to figure out in their head the order when the list is displayed in Rotation Order.
      isSeniorityOverridden:
        type: boolean
        description: True if the Seniority for the piece of equipment was manually overridden. Set if a user has gone in and explicitly updated the seniority base information. Indicates that underlying numbers were manually overridden.
      seniorityOverrideReason:
        type: string
        maxLength: 2048
        description: A text reason for why the piece of equipments underlying data was overridden to change their seniority number.
      seniorityEffectiveDate:
        type: string
        format: date-time
        description: The time the seniority data in the record went into effect. Used to populate the SeniorityAudit table when the seniority data is next updated.
      yearsOfService:
        type: number
        format: float
        description: The number of years of active service of this piece of equipment at the time seniority is calculated - April 1 of the current FY.
      serviceHoursLastYear:
        type: number
        format: float
        description: Number of hours of service by this piece of equipment in the previous fiscal year
      serviceHoursTwoYearsAgo:
        type: number
        format: float
        description: Number of hours of service by this piece of equipment in the fiscal year before the last one - e.g. if current year is FY2018 then hours in FY2016
      serviceHoursThreeYearsAgo:
        type: number
        format: float
        description: Number of hours of service by this piece of equipment in the fiscal year three years ago - e.g. if current year is FY2018 then hours in FY2015
      archiveCode:
        type: string
        maxLength: 50
        description: TO BE REVIEWED - A reason code indicating why a piece of equipment has been archived.
      archiveReason:
        type: string
        maxLength: 2048
        description: An optional comment about why this piece of equipment has been archived.
      archiveDate:
        type: string
        format: date-time
        description: The date on which a user most recenly marked this piece of equipment as archived.
      dumpTruck:
        $ref: '#/definitions/DumpTruck'
        description: A link to a dump truck set if this piece of equipment is an equipment type flagged as a dump truck.
      equipmentAttachments:
        type: array
        items:
          $ref: '#/definitions/EquipmentAttachment'
      notes:
        type: array
        items:
          $ref: '#/definitions/Note'
      attachments:
        type: array
        items:
          $ref: '#/definitions/Attachment'
      history:
        type: array
        items:
          $ref: '#/definitions/History'
      seniorityAudit:
        type: array
        items:
          $ref: '#/definitions/SeniorityAudit'
...
  EquipmentType:
    type: object
    description: A global and local area list of types of equipment for which a request can be generated and for which owners can register their equipment. If the local area is NULL, the type is visible Province-wide. In the previous instance of this application, all equipment types were assigned to the local area and transitioning to a province-wide model is being attempted (and may not work for the end users). An attempt to achieve a &quot;mostly province-wide&quot; list is being made during the data conversion process - mapping local area equipment types to provincial or local area types as part of data conversion.
    required:
      - id
    properties:
      id:
        format: int32
        type: integer
        description: A system-generated unique identifier for an EquipmentType
      localArea:
        $ref: '#/definitions/LocalArea'
      name:
        type: string
        maxLength: 50
        description: A human-friendly name for the Equipment Type that is displayed on screens with limited screen real estate available.
      description:
        type: string
        maxLength: 2048
        description: A extended description of the equipment type.
      equipRentalRateNo:
        type: number
        format: float
        description: TO BE REVIEWED WITH THE BUSINESS - WHAT IS THIS?
      equipRentalRatePage:
        type: number
        format: float
        description: TO BE REVIEWED WITH THE BUSINESS - WHAT IS THIS?
      maxHours:
        type: number
        format: float
        description: TO BE REVIEWED WITH THE BUSINESS - WHAT IS THIS?
      extendHours:
        type: number
        format: float
        description: TO BE REVIEWED WITH THE BUSINESS - WHAT IS THIS?
      maxHoursSub:
        type: number
        format: float
        description: TO BE REVIEWED WITH THE BUSINESS - WHAT IS THIS?
      blocks:
        format: int32
        type: integer
        description: The number of blocks defined for this equipment type. Currently, either 2 for Dump Truck equipment and 1 for non-Dump Truck equipment. Unlikely to change, particularly without a significant impact on code.
...
  EquipmentAttachment:
    type: object
    description: An Equipment Attachment associated with a piece of Equipment.
    required:
      - id
      - attachment
    properties:
      id:
        format: int32
        type: integer
        description: A system-generated unique identifier for an EquipmentAttachment
      equipment:
        $ref: '#/definitions/Equipment'
      attachment:
        type: string
        maxLength: 255
        description: The attachment type as entered by the user (per the business - no lookup).
      description:
        type: string
        maxLength: 2048
        description: A description of the equipment attachment if the Equipment Attachment Type is insufficient.
...
  Contact:
    type: object
    description: A person and their related contact information linked to one or more entities in the system. For examples, there are contacts for Owners, Projects.
    required:
      - id
    properties:
      id:
        format: int32
        type: integer
        uniqueItems: true
        description: A system-generated unique identifier for a Contact
      givenName:
        type: string
        maxLength: 50
        description: The given name of the contact.
      surname:
        type: string
        maxLength: 50
        description: The surname of the contact.
      organizationName:
        type: string
        maxLength: 150
        description: The organization name of the contact.
      role:
        type: string
        maxLength: 100
        description: The role of the contact. UI controlled as to whether it is free form or selected from an enumerated list - for initial implementation, the field is freeform.
      notes:
        type: string
        maxLength: 150
        description: A note about the contact maintained by the users.
      emailAddress:
        type: string
        maxLength: 255
        description: The email address for the contact.
      workPhoneNumber:
        type: string
        maxLength: 20
        description: The work phone number for the contact.
      mobilePhoneNumber:
        type: string
        maxLength: 20
        description: The mobile phone number for the contact.
      faxPhoneNumber:
        type: string
        maxLength: 20
        description: The fax phone number for the contact.
      address1:
        type: string
        maxLength: 80
        description: Address 1 line of the address.
      address2:
        type: string
        maxLength: 80
        description: Address 2 line of the address.
      city:
        type: string
        maxLength: 100
        description: The City of the address.
      province:
        type: string
        maxLength: 50
        description: The Province of the address.
      postalCode:
        type: string
        maxLength: 15
        description: The postal code of the address.
  Note:
    type: object
    description: Text entered about an entity in the application - e.g. piece of Equipment, an Owner, a Project and so on.
    required:
      - id
      - text
    properties:
      id:
        format: int32
        type: integer
        description: A system-generated unique identifier for a Note
      text:
        type: string
        maxLength: 2048
        description: Notes entered by users about instance of entities - e.g. School Buses and School Bus Owners
      isNoLongerRelevant:
        type: boolean
        description: A user set flag that the note is no longer relevant. Allows the note to be retained for historical reasons,  but identified to other users as no longer applicable.
  Attachment:
    type: object
    description: Uploaded documents related to entity in the application - e.g. piece of Equipment, an Owner, a Project and so on.
    required:
      - id
    properties:
      id:
        format: int32
        type: integer
        uniqueItems: true
        description: A system-generated unique identifier for an Attachment
      fileName:
        type: string
        maxLength: 2048
        description: Filename as passed by the user uploading the file
      fileContents:
        type: string
        format: binary
        description: Binary contents of the file
      description:
        type: string
        maxLength: 2048
        description: A note about the attachment,  optionally maintained by the user.
      type:
        type: string
        maxLength: 255
        description: Type of attachment
  History:
    type: object
    description: A log entry created by the system based on a triggering event and related to an entity in the application - e.g. piece of Equipment, an Owner, a Project and so on.
    required:
      - id
      - historyText
    properties:
      id:
        format: int32
        type: integer
        description: A system-generated unique identifier for a History
      historyText:
        type: string
        maxLength: 2048
        description: The text of the history entry tracked against the related entity.
      createdDate:
        type: string
        format: date-time
        description: Date the record is created.
...
  ServiceArea:
    type: object
    description: The Ministry of Transportation and Infrastructure SERVICE AREA.
    required:
      - id
      - ministryServiceAreaID
      - name
      - district
      - startDate
    properties:
      id:
        format: int32
        type: integer
        uniqueItems: true
        description: A system-generated unique identifier for a ServiceArea
      ministryServiceAreaID:
        format: int32
        type: integer
        description: A system generated unique identifier. NOT GENERATED IN THIS SYSTEM.
      areaNumber:
        format: int32
        type: integer
        description: A number that uniquely defines a Ministry Service Area.
      name:
        type: string
        maxLength: 150
        description: The Name of a Ministry Service Area.
      district:
        $ref: '#/definitions/District'
        description: The district in which the Service Area is found.
      startDate:
        type: string
        format: date-time
        description: The DATE the business information came into effect. - NOT CURRENTLY ENFORCED IN THIS SYSTEM
      endDate:
        type: string
        format: date-time
        description: The DATE the business information ceased to be in effect. - NOT CURRENTLY ENFORCED IN THIS SYSTEM
  LocalArea:
    type: object
    description: A HETS-application defined area that is within a Service Area.
    required:
      - id
      - startDate
    properties:
      id:
        format: int32
        type: integer
        description: A system-generated unique identifier for a LocalArea
      localAreaNumber:
        format: int32
        type: integer
        description: A system-generated, visible to the user number for the Local Area
      name:
        type: string
        maxLength: 150
        description: The full name of the Local Area
      serviceArea:
        $ref: '#/definitions/ServiceArea'
        description: The Service Area in which the Local Area is found.
      startDate:
        type: string
        format: date-time
        description: The DATE the business information came into effect. - NOT CURRENTLY ENFORCED IN THIS SYSTEM
      endDate:
        type: string
        format: date-time
        description: The DATE the business information ceased to be in effect. - NOT CURRENTLY ENFORCED IN THIS SYSTEM
...
  ProjectSearchResultViewModel:
    type: object
    x-codegen-viewmodel: true
    required:
      - id
    properties:
      id:
        format: int32
        type: integer
        x-type-options:
          min: 1
          max: 1000
      localArea:
        $ref: '#/definitions/LocalArea'
      name:
        type: string
      primaryContact:
        $ref: '#/definitions/Contact'
      hires:
        format: int32
        type: integer
        description: count of RentalAgreement.status is Active for the project
      requests:
        format: int32
        type: integer
        description: count of RentalRequest.status is Active for the project
      status:
        type: string
        description: Project status        
...
securityDefinitions: {}